backend/requirements.txt
django==4.2.16
djangorestframework==3.15.2
channels==4.0.0
daphne==4.0.0
django-cors-headers==4.3.1
channels-redis==4.2.0
mediapipe==0.10.14
opencv-python-headless==4.10.0.84
tensorflow==2.16.2
numpy==1.26.4
Pillow==10.4.0
pyttsx3==2.90
gTTS==2.5.1
python-dotenv==1.0.1
whitenoise==6.7.0
gunicorn==22.0.0
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2
